### 2019-03-06 #1304 @sauln
Sorry the paper was hidden in another branch. It's on master now.
