### 2019-03-06 #1304 @sauln
Thank you @whedon.  For possible reviewers, @corybrunson may be interested, as well as @rrrlw or @lmcinnes.

To review some of the visualization aspects of the software, @a-slide, @zfrenchee, or @ixjlyons, or @ejolly may be interested.

Thank you!
