### 2019-03-07 3230172
* aimed to emphasize the need for mapper in TDA, and kepler-mapper for mapper * added links to package libraries * adding demo image
