+++ paper.md
@@ -29,21 +29,22 @@ bibliography: paper.bib
 
 # Summary
 
-Topological data analysis (TDA) is a field analysis focused on understanding the shape and structure of data. By computing topological descriptors of data such as connected components, loops, and voids, we are better able to find hidden relationships [@edelsbrunner2010computational], [@carlsson2009topology]. Mapper is one main technique from the field that is designed for visualization of high dimensional topological structures.
+Topological data analysis (TDA) is a field analysis focused on understanding the shape and structure of complex data. By computing topological descriptors of data such as connected components, loops, and voids, we are better able to find hidden relationships among noisy and high-dimensionality data [@edelsbrunner2010computational], [@carlsson2009topology]. However, raw TDA metrics cannot be readily visualized. To address this gap, [@Singh2007] developed the Mapper algorithm in order to facilitate graphical exploration of data topoloigcal structures. The work of [@Lumetal2013] sparked widespread interest in the Mapper technique by demonstrating its use in multiple domains, including political science, biology, and sports analytics.
 
-Kepler Mapper employs approaches based on the Mapper algorithm (Singh et al.) as first described in the paper “Topological Methods for the Analysis of High Dimensional Data Sets and 3D Object Recognition” [@Singh2007].
-The work of [@Lumetal2013] sparked widespread interest in the technique by demonstrating the use in multiple domains, such as political science, biology, and sports analytics.
+This library, Kepler Mapper, is a Python implementation of the Mapper algorithm as first described in the paper “Topological Methods for the Analysis of High Dimensional Data Sets and 3D Object Recognition” [@Singh2007]. Kepler Mapper presents an intuitive interface for the Mapper algorithm along with multiple comprehension methods for visualizing the network graph that Mapper produces.
+We leverage Scikit-Learn API-compatible cluster and scaling algorithms to construct network graphs in a flexible and user-friendly way.
+We also an provide extensive suite of tutorials detailing the use of Kepler Mapper for simple and complex use cases.
 
-This library presents an intuitive interface for Mapper algorithm along with multiple comprehension methods for visualizing the construction.
-We leverage Scikit-Learn API-compatible cluster and scaling algorithms to construct Mappers in a flexible and user-friendly way.
-We also an provide extensive suite of tutorials detailing the use of Mapper for simple and complex use cases.
+![breast-cancer](http://i.imgur.com/ewjRodK.png)
+
+_Figure -- Example Kepler Mapper graph visualization using the [Wisconsin Breast Cancer Dataset](https://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Diagnostic))_
 
 
 # Library Details
 
-Kepler Mapper provides an object-oriented API for constructing lenses and building Mapper. The module employs the strategy pattern giving users control over clustering algorithm, covering scheme, and nerve scheme. This allows the module to be flexible for many use cases. Clustering strategies follow the Scikit-Learn clusterer interface [@scikit-learn]. We have developed similar interfaces for `Cover` classes and `Nerve` classes and provided default implementations that are most commonly found in the literature.
+Kepler Mapper provides an object-oriented API for constructing lenses and building Mapper. The module employs the [strategy pattern](https://en.wikipedia.org/wiki/Strategy_pattern), giving users control over clustering algorithm, covering scheme, and nerve scheme. This allows the module to be flexible for many use cases. Clustering strategies follow the Scikit-Learn clusterer interface [@scikit-learn]. We provide similar interfaces for `Cover` classes and `Nerve` classes as well as default implementations of those classes that are most commonly found in the literature.
 
[…]
