### 2019-03-09 #1304 @leouieda
@lmcinnes and @ixjlyons would be interested in reviewing this submission for JOSS?
