### 2019-03-11 #1304 @leouieda
@lmcinnes as long as the authors are OK with the wait, it's fine by me.
