### 2019-03-11 #1315 @whedon
[REVIEW]: Kepler Mapper: A flexible Python implementation of the Mapper algorithm.
**Submitting author:** @sauln (<a href="http://orcid.org/0000-0002-8549-9810">Nathaniel Saul</a>)
**Repository:** <a href="https://github.com/scikit-tda/kepler-mapper" target ="_blank">https://github.com/scikit-tda/kepler-mapper</a>
**Version:** 1.3.3
