### 2019-03-11 #1315 @leouieda
@lmcinnes and @ixjlyons please read the reviewer guidelines (if you haven't already) and fill out the review checklist. If you open any issues or PRs in the software repository, please add `openjournals/joss-reviews#1315` to the description so we can keep track of the process.

Please feel free to ping me at any time :bellhop_bell:
