### 2019-03-11 scikit-tda/kepler-mapper#147 @leouieda
Affiliation missing for Sam Mangham (JOSS paper)
The affiliation index for Sam Mangham is marked as `4` but there is no definition of an index 4 affiliation below it.

openjournals/joss-reviews#1315
