### 2019-03-11 scikit-tda/kepler-mapper#147 @sauln
@smangham What would you like your affiliation to be?
