### 2019-03-12 scikit-tda/kepler-mapper#147 @smangham
Could my affiliation be 'Department of Electronics & Computer Science, University of Southampton, Southampton, SO17 1BJ, UK' please! My paper name is usually 'S. W. Mangham' or 'Sam W. Mangham', whichever fits your style.
