### 2019-03-13 56ca48d
move pull-in of js and css file into jinja template
