### 2019-03-14 143b6fd
move tooltip rendering into single render_template call
