### 2019-03-14 4a70bbb
update test to not use env
