### 2019-03-14 dbb4b75
add row_color_function and node_color_function
