### 2019-03-15 0446db6
add default row and node color functions which can be imported by tests etc
