### 2019-03-15 074b4aa
update plotly to use new color data and function refactor
