### 2019-03-15 9d2f609
update visuals test for new color approach
