### 2019-03-15 3208e4e
Update name and affiliation for Sam Mangham
Fixes #147 

openjournals/joss-reviews#1315
