### 2019-03-15 601834d
Typo fix in  Contributing document
