### 2019-03-15 0af3bc1
remove code of conduct from contributing page
