### 2019-03-15 a3446ec
Fix figure caption formatting for JOSS paper
The formatting for this is a bit odd but `![caption here](file.png)`

openjournals/joss-reviews#1315
