### 2019-03-15 #1315 @leouieda
@sauln one more fix to the paper formatting and we should be good to go (scikit-tda/kepler-mapper#155)
