### 2019-03-16 scikit-tda/kepler-mapper#156 @leouieda
Replace link to data with citation in JOSS paper
The Wisconsin Breast Cancer Dataset [website](https://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Diagnostic)) has a [citation policy](https://archive.ics.uci.edu/ml/citation_policy.html) asking to cite it using:

```
@misc{Dua:2019 ,
author = "Dua, Dheeru and Graff, Casey",
year = "2017",
title = "{UCI} Machine Learning Repository",
url = "http://archive.ics.uci.edu/ml",
institution = "University of California, Irvine, School of Information and Computer Sciences" }
```

Since Whedon is complaining about the link in the caption of the JOSS paper figure, it might be better to replace it with the citation instead.

openjournals/joss-reviews#1315
