### 2019-03-18 33bd466
whitespace control on rendered templates
