+++ kmapper/visuals.py
@@ -6,6 +6,7 @@ import os
 from collections import defaultdict
 from ast import literal_eval
 from jinja2 import Environment, FileSystemLoader, Template, StrictUndefined
+from css_html_js_minify import html_minify, js_minify, css_minify
 
 
 colorscale_default = [
@@ -280,6 +281,19 @@ def render_template(
             trim_blocks=True,
             undefined=StrictUndefined,
             )
+    
+    static_dir = os.path.join(module_root, 'static')
+    
+    # Find the absolute module path and the static files
+    js_path = os.path.join(static_dir, "kmapper.js")
+    with open(js_path, "r") as f:
+        js_text = f.read()
+    js_text = js_minify(js_text)
+
+    css_path = os.path.join(static_dir, "style.css")
+    with open(css_path, "r") as f:
+        css_text = f.read()
+    css_text = css_minify(css_text)
 
     # Render the Jinja template, filling fields as appropriate
     rendered_template = env.get_template("base.html").render(
@@ -288,7 +302,9 @@ def render_template(
         histogram=histogram,
         dist_label="Node",
         mapper_data=mapper_data,
-        colorscale=colorscale
+        colorscale=colorscale,
+        js_text=js_text,
+        css_text=css_text,
     )
     
     return rendered_template
