### 2019-03-22 c082139
nodes map was useless
