### 2019-03-22 0599e9f
* move svg defs into html, because why were they in javascript * focus node pulses * register custom drag event so that we get the dragstart and dragend events; make cursor set check dependent on drag event flag. custom-register mouseover node position-fixing * implement layout freezing and unfreezing; add to help toolbar
