+++ kmapper/templates/static/kmapper.js
@@ -7,7 +7,7 @@ var w = width;
 var h = height;
 var padding = 40;
 
-var focus_node = null, highlight_node = null;
+var focus_node = null;
 var text_center = false;
 var outline = false;
 
@@ -17,8 +17,6 @@ var size = d3.scale.pow().exponent(1)
            .range([8,24]);
 
 // Variety of variable inits
-var highlight_color = "blue";
-var highlight_trans = 0.1;
 var default_node_color = "#ccc";
 var default_node_color = "rgba(160,160,160, 0.5)";
 var default_link_color = "rgba(160,160,160, 0.5)";
@@ -42,7 +40,7 @@ if (outline) {
 
 
 // We draw the graph in SVG
-var svg = d3.select("#canvas").append("svg")
+var svg = d3.select("#canvas svg")
           .attr("width", width)
           .attr("height", height);
 
@@ -129,6 +127,19 @@ var force = d3.layout.force()
             .charge(-1200)
             .size([w,h]);
 
+var dragging = false;
+
+var drag = force.drag()
+  .on("dragstart", function(d){
+    svg.style('cursor','grabbing');
+    d.fixed = true;
+    dragging = true;
+  })
+  .on('dragend', function(d){
+    dragging = false;
+  })
+  ;
+
 force
   .nodes(graph.nodes)
   .links(graph.links)
@@ -146,7 +157,8 @@ var node = g.selectAll(".node")
             .data(graph.nodes)
             .enter().append("g")
               .attr("class", "node")
-              .call(force.drag);
+              .attr("id", function(d){ return "node-" + d.name })
+              .call(drag);
 
 // Double clicking on a node will center on it.
 node.on("dblclick.zoom", function(d) { d3.event.stopPropagation();
@@ -155,40 +167,6 @@ node.on("dblclick.zoom", function(d) { d3.event.stopPropagation();
   zoom.translate([dcx,dcy]);
   g.attr("transform", "translate("+ dcx + "," + dcy  + ")scale(" + zoom.scale() + ")");
 });
-
-
-
-
-// Drop-shadow Filter
-var svg = d3.select("svg");
-var defs = svg.append("defs");
-var dropShadowFilter = defs.append('svg:filter')
-  .attr('id', 'drop-shadow')
-  .attr('filterUnits', "userSpaceOnUse")
-  .attr('width', '250%')
-  .attr('height', '250%');
-dropShadowFilter.append('svg:feGaussianBlur')
-  .attr('in', 'SourceGraphic')
-  .attr('stdDeviation', 12)
-  .attr('result', 'blur-out');
-dropShadowFilter.append('svg:feColorMatrix')
-  .attr('in', 'blur-out')
-  .attr('type', 'hueRotate')
-  .attr('values', 0)
-  .attr('result', 'color-out');
-dropShadowFilter.append('svg:feOffset')
-  .attr('in', 'color-out')
-  .attr('dx', 0)
-  .attr('dy', 0)
-  .attr('result', 'the-shadow');
-dropShadowFilter.append('svg:feComponentTransfer')
-  .attr('type', 'linear')
-  .attr('slope', 0.2)
-  .attr('result', 'shadow-opacity');
-dropShadowFilter.append('svg:feBlend')
-  .attr('in', 'SourceGraphic')
-  .attr('in2', 'the-shadow')
-  .attr('mode', 'normal');
 
 // Draw circles
 var circle = node.append("path")
@@ -201,7 +179,6 @@ var circle = node.append("path")
     console.log("becomes color ", color(d.color));
     return color(d.color);
   });
-//.style("filter", "url(#drop-shadow)");
 
 
[…]
