### 2019-03-29 #1315 @lmcinnes
I can start working on it next week once I have SciPy2019 commitments out of the way.
