### 2019-04-01 c2ceee1
remove minification because python 2.7 incompat
