### 2019-04-01 776cafd
fix remove_duplicate test after join char change
