### 2019-04-04 8a2a955
add release 1.2.0
