+++ RELEASE.txt
@@ -1,5 +1,9 @@
 Alpha
 
+1.2.0 (Feb 18, 2019)
+	New implementation of the cover API makes it consistent with the literature (your % overlap will probably have to be decreased when updating).
+	New documentation website (kepler-mapper.scikit-tda.org).
+
 1.1.6 (Nov 8, 2018)
 	Plotly visualization interface.
 	Networkx adapter
