### 2019-04-05 #1315 @ixjlyons
My review:

The purpose of Kepler Mapper is clear from the docs and paper, and it appears to me to fulfill that purpose. I appreciate the background page of the docs as well as the applications page. The tutorials take the documentation above and beyond a reference for library usage and look nice overall. The code seems to be structured sensibly and it looks like the library is quite flexible/extensible. The paper is clear and concise.

I think the examples page could use a little work. As it is now, there isn't really any information about the examples and there isn't really a connection to the source code. Something like [sphinx-gallery](https://sphinx-gallery.github.io/) would work well, especially if you could serve and link to the interactive HTML interface. Some information about the examples themselves would also be helpful, such as a description of the dataset and maybe some insight Mapper provides.

There are [mutable default arguments](https://docs.python-guide.org/writing/gotchas/) in a handful of places. I didn't check thoroughly, but it appears to me that those objects are used in such a way that there shouldn't be a problem, and it may be totally intentional. I just wanted to bring it up to make sure the potential pitfalls were considered and avoided.

Some minor things I noticed that didn't seem to warrant their own issues:

- Typo in the paper, second paragraph: "We also an provide extensive..." -> "We also provide an extensive..."
- I think there may be some missing DOIs in the references.
- I'd consider hiding `SimplicialNerve` until it's implemented.
- The last example in the `KeplerMapper.project()` docstring should probably be moved/copied to the `fit_transform()` docstring.

Overall, Kepler Mapper appears to be in great shape. I recommend accepting.
