### 2019-04-06 scikit-tda/kepler-mapper#164 @sauln
Use sphinx-gallery for doc examples
The [documentation examples](https://kepler-mapper.scikit-tda.org/examples.html) are currently just a few images while the examples in the [examples folder](https://github.com/scikit-tda/kepler-mapper/tree/master/examples) provide executable code. We should be able to use [sphinx-gallery](https://sphinx-gallery.github.io) to run the code in the example scripts and display the results.

This will look much nicer than the current system and will address suggestions in openjournals/joss-reviews#1315
