### 2019-04-06 scikit-tda/kepler-mapper#165 @sauln
Remove mutable arguments
As mentioned in openjournals/joss-reviews#1315, some functions have mutable arguments.

I found the following instances of this:
https://github.com/scikit-tda/kepler-mapper/blob/8a2a955e2b7302ee3c37e5c6a2c6ebfc1924a39b/kmapper/kmapper.py#L630-L632

These could easily be transformed into something like
```
def f(param=None):
    param = [] if param is None else param
```
