### 2019-04-06 scikit-tda/kepler-mapper#166 @sauln
JOSS Review Suggested Edits
Review in openjournals/joss-reviews#1315 by @ixjlyons suggested a few minor fixes along with #164 and #165.

# JOSS Paper:
- [x] Typo in the paper, second paragraph: "We also an provide extensive..." -> "We also provide an extensive..."
- [x] I think there may be some missing DOIs in the references.

# KM implementation:
- [x] I'd consider hiding `SimplicialNerve` until it's implemented.
- [x] The last example in the `KeplerMapper.project()` docstring should probably be moved/copied to the `fit_transform()` docstring.
