### 2019-04-07 scikit-tda/kepler-mapper#165 @cmorph1
Hi @sauln
I'm new to os, but I'd like to take a look at this
