### 2019-04-07 scikit-tda/kepler-mapper#165 @sauln
Yeah, absolutely! Let me know how I can support you!
