### 2019-04-07 scikit-tda/kepler-mapper#165 @cmorph1
Excellent, leave this with me. I'll come back to you if I get stuck.
