### 2019-04-08 b73fdfa
building out sphinx gallery
