+++ examples/README.txt
@@ -1,0 +1,4 @@
+KM Examples Gallery
+=======================
+
+Below is a gallery of examples

[…]
