### 2019-04-08 cb3d80f
added a test and updated docstring and imports
