### 2019-04-08 303e1ff
Include random state in tests
