### 2019-04-08 bfa944c
Add note for PR #163
