### 2019-04-08 23c87e7
mutable arguments removed (not including color_default dict)
