### 2019-04-08 scikit-tda/kepler-mapper#165 @leouieda
I would add to be wary of dictionaries and non-builtin types (like instances of your own classes) as well.
