### 2019-04-08 #1315 @leouieda
@ixjlyons thank you for your review and suggestions :clap:

@sauln I see you've opened issues for the points raised by @ixjlyons (thank you for doing that). Please try to address these points, in particular the use of mutable defaults. This is a dangerous pitfall in Python that lead to extremely hard to catch bugs (I speak from experience).
