### 2019-04-09 0100f87
Added lens name test
