### 2019-04-22 #1315 @lmcinnes
Sorry for being so slow. Once I got the time this went very quickly because everything was mostly very well laid out and easy to test and check. I agree with @ixjlyons that a sphinx-gallery style approach to the examples might be nice (and hopefully not that hard to set up).

You can also add DOIs for a few of the references. Carlssons paper is hDOI: 10.1090/S0273-0979-09-01249-X, and Edelsbrunner is DOI: 10.1007/978-3-540-33259-6_7.

Other than that it all meets the requirements and looks good to me.
