### 2019-04-24 1f1833c
fixed test and removed mutable default_colorscale from arguments
