### 2019-04-24 scikit-tda/kepler-mapper#165 @sauln
#167 fixes most of the occurrences of mutable arguments. I can't find any dictionaries and all that is left is some non-builtin types.

I'm hesitant to change the instances when it's our class or a sklearn class. For one, I like how [they are rendered in the documentation](https://kepler-mapper.scikit-tda.org/reference/stubs/kmapper.KeplerMapper.html#kmapper.KeplerMapper.map) and I'm not sure how to best convey the defaults when `cover=None`. They also all have a `fit` method that is called within the function, so we can be sure the objects aren't being fit to the wrong data.

@leouieda, if you're steadfast on these non-builtin types, we can figure out how to remove them without compromising the documentation.
