### 2019-04-24 6dd07ed
add colorscale test
