### 2019-04-24 36a9a9b
move fit_transform example to fit_transform docstring
