### 2019-04-24 2a649ea
Generated gh-pages for commit d4ed39f Merge: ebe6bda 6dd07ed Author: Nathaniel Saul <nat@saulgill.com>
    Merge pull request #167 from cmorph1/remove-mutable-arguments
