### 2019-04-30 #1315 @leouieda
@lmcinnes thank you for your feedback!

 @sauln both reviews are very positive and there only a few minor things that need to be implemented. Please let us know when you've had time to implement the suggestions of both reviewers. To make it easier to track, here is a checklist (feel free to tick things off as you go):

- [ ] More contextual information for the examples
- [ ] Add missing DOIs
- [ ] Replace mutable default arguments
- [ ] Hide SimplicialNerve until it's implemented
- [ ] last example in the KeplerMapper.project() docstring should probably be moved/copied to the fit_transform() docstring
- [ ] Use sphinx-gallery for the documentation (optional but nice to have)

Take your time with these and please let us know if you want any clarifications or further input.
