### 2019-04-30 scikit-tda/kepler-mapper#165 @leouieda
@sauln I understand your points and have struggled with that myself. The main problem is that the default arguments are instantiated at import time. So calling `fit` the default argument will have side effects on other instances. For example:

```python
class Estimator:
    def __init__(self):
        self.params_ = None
    def fit(self, X, y):
        self.params_ = [X, y]
        return self

class OtherEstimator:
    def __init__(self, estimator=Estimator()):
        self.estimator = estimator
    def fit(self, X, y):
        self.estimator.fit(X, y)
        return self

a = OtherEstimator()
b = OtherEstimator()

a.fit(1, 2)

a.estimator is b.estimator
# True

print(a.estimator.params_)
# [1, 2]

print(b.estimator.params_)
# [1, 2]
```

This can lead to nasty bugs. It's convenient for the docs but I don't think the trade-off is worth it.

If you look at the scikit-learn defaults and other core Python libraries, you'll see some different strategies for dealing with this. Many packages default to `None` if want an empty container or a string if there are many options with names. See `cmap="viridis"` in matplotlib or the [`cv` argument in `cross_val_score`](https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.cross_val_score.html#sklearn.model_selection.cross_val_score).
That is certainly the approach I'd recommend.
