### 2019-05-02 #1315 @sauln
Thank you @leouieda, @lmcinnes, and @ixjlyons for taking the time to edit and review.  The feedback is encouraging and we are working on each of the changes that have been suggested.

@leouieda, the past few weeks have been crazy for me, but I am working on the examples and mutable defaults currently. I hope to finish up within the next 2 weeks.

I've created the checklist here, as github won't let me check your boxes.

- [ ] More contextual information for the examples
- [x] Add missing DOIs
- [ ] Replace mutable default arguments
- [x] Hide SimplicialNerve until it's implemented
- [x]  last example in the KeplerMapper.project() docstring should probably be moved/copied to the fit_transform() docstring
- [ ]   Use sphinx-gallery for the documentation (optional but nice to have)
