+++ Makefile
@@ -9,4 +9,4 @@ gh-pages:
 	mv -fv docs/_build/html/* .
 	rm -rf docs
 	git add -A
-	git commit -m "Generated gh-pages for `git log master -1 --pretty=short --abbrev-commit`" && git push scikit gh-pages ; git checkout master
+	git commit -m "Generated gh-pages for `git log master -1 --pretty=short --abbrev-commit`" && git push origin gh-pages ; git checkout master
