+++ _sources/generated/gallery/plot_breast_cancer.rst.txt
@@ -1,269 +1,0 @@
-.. note::
-    :class: sphx-glr-download-link-note
-
-    Click :ref:`here <sphx_glr_download_generated_gallery_plot_breast_cancer.py>` to download the full example code
-.. rst-class:: sphx-glr-example-title
-
-.. _sphx_glr_generated_gallery_plot_breast_cancer.py:
-
-
-Breast Cancer
-================
-
-
-
-This example generates a Mapper built from the `Wisconsin Breast Cancer Dataset <https://www.kaggle.com/uciml/breast-cancer-wisconsin-data>`_.
-
-
-.. image:: ../../../examples/images/breast-cancer.png
-
-
-
-.. rst-class:: sphx-glr-script-out
-
- Out:
-
- .. code-block:: none
-
-    KeplerMapper(verbose=3)
-    ..Composing projection pipeline of length 1:
-            Projections: l2norm
-            Distance matrices: False
-            Scalers: MinMaxScaler(copy=True, feature_range=(0, 1))
-    ..Projecting on data shaped (569, 31)
-
-    ..Projecting data using: l2norm
-
-    ..Scaling with: MinMaxScaler(copy=True, feature_range=(0, 1))
-
-    Mapping on data shaped (569, 31) using lens shaped (569, 2)
-
-    Minimal points in hypercube before clustering: 2
-    Creating 225 hypercubes.
-    Cube_0 is empty.
-
-    Cube_1 is empty.
-
-    Cube_2 is empty.
-
-    Cube_3 is empty.
-
-    Cube_4 is empty.
-
-    Cube_5 is empty.
-
-    Cube_6 is empty.
-
-       > Found 2 clusters in hypercube 7.
-    Cube_8 is empty.
-
-    Cube_9 is empty.
-
-       > Found 2 clusters in hypercube 10.
-       > Found 2 clusters in hypercube 11.
-    Cube_12 is empty.
-
-    Cube_13 is empty.
-
-    Cube_14 is empty.
-
-       > Found 2 clusters in hypercube 15.
-    Cube_16 is empty.
-
-    Cube_17 is empty.
-
-    Cube_18 is empty.
-
-    Cube_19 is empty.
-
-    Cube_20 is empty.
-
-       > Found 2 clusters in hypercube 21.
-       > Found 2 clusters in hypercube 22.
-    Cube_23 is empty.
-
-       > Found 2 clusters in hypercube 24.
-    Cube_25 is empty.
-
-       > Found 2 clusters in hypercube 26.
-       > Found 2 clusters in hypercube 27.
-    Cube_28 is empty.
-
-       > Found 2 clusters in hypercube 29.
-       > Found 2 clusters in hypercube 30.
-       > Found 2 clusters in hypercube 31.
-       > Found 2 clusters in hypercube 32.
-       > Found 2 clusters in hypercube 33.
-       > Found 2 clusters in hypercube 34.
-       > Found 2 clusters in hypercube 35.
-       > Found 2 clusters in hypercube 36.
-    Cube_37 is empty.
-
-       > Found 2 clusters in hypercube 38.
-       > Found 2 clusters in hypercube 39.
-    Cube_40 is empty.
-
-       > Found 2 clusters in hypercube 41.
-    Cube_42 is empty.
-
-       > Found 2 clusters in hypercube 43.
-       > Found 2 clusters in hypercube 44.
-       > Found 2 clusters in hypercube 45.
-       > Found 2 clusters in hypercube 46.
-    Cube_47 is empty.
-
-    Cube_48 is empty.
-
-       > Found 2 clusters in hypercube 49.
-    Cube_50 is empty.
-
-    Cube_51 is empty.
-
[…]
