+++ _sources/generated/gallery/sg_execution_times.rst.txt
@@ -1,12 +1,0 @@
-
-:orphan:
-
-.. _sphx_glr_generated_gallery_sg_execution_times:
-
-Computation times
-=================
-**00:01.304** total execution time for **generated_gallery** files:
-
-- **00:01.304**: :ref:`sphx_glr_generated_gallery_plot_breast_cancer.py` (``plot_breast_cancer.py``)
-- **00:00.000**: :ref:`sphx_glr_generated_gallery_plot_cat.py` (``plot_cat.py``)
-- **00:00.000**: :ref:`sphx_glr_generated_gallery_plot_horse.py` (``plot_horse.py``)
