+++ _sources/reference/stubs/kmapper.SimplicialNerve.rst.txt
@@ -1,23 +1,0 @@
-kmapper.SimplicialNerve
-=======================
-
-.. currentmodule:: kmapper
-
-.. autoclass:: SimplicialNerve
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~SimplicialNerve.__init__
-      ~SimplicialNerve.compute
-   
-   
-
-   
-   
-   

[…]
