### 2019-05-18 deb847f
functional sphinx-gallery with html output linking
