+++ examples/plot_cat.py
@@ -6,6 +6,9 @@
 
 This example generates a Mapper built from a point-cloud sampled from a 3D model of a cat.
 
+`Visualization of the cat mapper <../../_static/cat.html>`_
+
+
 """
 
 
@@ -17,7 +20,6 @@ data = np.genfromtxt('data/cat-reference.csv', delimiter=',')
 
 mapper = km.KeplerMapper(verbose=2)
 
-
 lens = mapper.fit_transform(data)
 
 graph = mapper.map(lens,
@@ -26,7 +28,7 @@ graph = mapper.map(lens,
                    cover=km.Cover(n_cubes=15, perc_overlap=0.2))
 
 mapper.visualize(graph,
-                 path_html="output/cat_keplermapper.html")
+                 path_html="output/cat.html")
 
 km.draw_matplotlib(graph)
 
