### 2019-05-18 0b0133d
add layout option for draw_matplotlib
