### 2019-05-18 448aeb1
fix latex and link to choosing a lens
