### 2019-05-18 a051b53
adding generated dir to docs
