+++ _sources/generated/gallery/plot_breast_cancer.rst.txt
@@ -1,0 +1,289 @@
+.. note::
+    :class: sphx-glr-download-link-note
+
+    Click :ref:`here <sphx_glr_download_generated_gallery_plot_breast_cancer.py>` to download the full example code
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_generated_gallery_plot_breast_cancer.py:
+
+
+Breast Cancer
+================
+
+
+
+This example generates a Mapper built from the `Wisconsin Breast Cancer Dataset <https://www.kaggle.com/uciml/breast-cancer-wisconsin-data>`_.
+
+`Visualization of the breast cancer mapper <../../_static/breast-cancer.html>`_
+
+
+The reasoning behind the choice of lenses in the demonstration above is:
+
+- **For lens1:** Lenses that make biological sense; in other words, lenses that highlight special features in the data, that I know about. 
+- **For lens2:** Lenses that disperse the data, as opposed to clustering many points together.
+
+In the case of this particualr data, using an anomaly score (in this case calculated using the IsolationForest from sklearn) makes biological sense since cancer cells are anomalous. For the second lens, we use the :math:`l^2` norm.
+
+For an interactive exploration of lens for the breast cancer, see the `Choosing a lens notebook <../../notebooks/Cancer-demo.html>`_.
+
+
+
+.. image:: ../../../examples/images/breast-cancer.png
+
+
+.. image:: /generated/gallery/images/sphx_glr_plot_breast_cancer_001.png
+    :class: sphx-glr-single-img
+
+
+.. rst-class:: sphx-glr-script-out
+
+ Out:
+
+ .. code-block:: none
+
+    KeplerMapper(verbose=3)
+    ..Composing projection pipeline of length 1:
+            Projections: l2norm
+            Distance matrices: False
+            Scalers: MinMaxScaler(copy=True, feature_range=(0, 1))
+    ..Projecting on data shaped (569, 31)
+
+    ..Projecting data using: l2norm
+
+    ..Scaling with: MinMaxScaler(copy=True, feature_range=(0, 1))
+
+    Mapping on data shaped (569, 31) using lens shaped (569, 2)
+
+    Minimal points in hypercube before clustering: 2
+    Creating 225 hypercubes.
+    Cube_0 is empty.
+
+    Cube_1 is empty.
+
+    Cube_2 is empty.
+
+    Cube_3 is empty.
+
+    Cube_4 is empty.
+
+    Cube_5 is empty.
+
+    Cube_6 is empty.
+
+       > Found 2 clusters in hypercube 7.
+    Cube_8 is empty.
+
+    Cube_9 is empty.
+
+       > Found 2 clusters in hypercube 10.
+       > Found 2 clusters in hypercube 11.
+    Cube_12 is empty.
+
+    Cube_13 is empty.
+
+    Cube_14 is empty.
+
+       > Found 2 clusters in hypercube 15.
+    Cube_16 is empty.
+
+    Cube_17 is empty.
+
+    Cube_18 is empty.
+
+    Cube_19 is empty.
+
+    Cube_20 is empty.
+
+       > Found 2 clusters in hypercube 21.
+       > Found 2 clusters in hypercube 22.
+    Cube_23 is empty.
+
+       > Found 2 clusters in hypercube 24.
+    Cube_25 is empty.
+
+       > Found 2 clusters in hypercube 26.
+       > Found 2 clusters in hypercube 27.
+    Cube_28 is empty.
+
+       > Found 2 clusters in hypercube 29.
+       > Found 2 clusters in hypercube 30.
+       > Found 2 clusters in hypercube 31.
+       > Found 2 clusters in hypercube 32.
+       > Found 2 clusters in hypercube 33.
[…]
