### 2019-05-22 #1315 @sauln
At least another week. I have a [draft of the sphinx-gallery](https://kepler-mapper.scikit-tda.org/generated/gallery/index.html) working but it needs a little more work before completion.

I've been convinced about the mutable defaults.  I think the best way will be to go with `cover=None` and then add the default rather than choosing a set of strings.
