### 2019-07-11 ac7d637
define path length
