### 2019-07-11 497f003
definition outline
