+++ kmapper/kmapper.py
@@ -898,14 +898,40 @@ class KeplerMapper(object):
         X = X_blend
 
         return X
-
+    
+    def path_length(
+        self,
+        X,
+        a,
+        b
+    ):
+        """Let members a, b each in nodes n, m. And define path length r of a and b; least number of edges that connect a, b. 
+           If a and b is connected by same node, then r is 0. And if a and b is not connected, then r is -1.
+        
+        Parameters
+        ----------
+        X : Numpy Array
+            The data mapped
+        a : a member
+        b : a member
+        
+        Returns
+        -------
+        r : int
+            path length r of a and b
+        
+        """
+        
+        
+        
+        return r
+        
+        
     def digitize_relationship(
         self,
         X,
     ):
-        """Let members a, b each in nodes n, m. And define path length r of a and b; least number of edges that connect a, b. 
-           If a and b is connected by same node, then r is 0. And if a and b is not connected, then r is -1.
-           This function print out Dataframe that every r of the data mapped.
+        """This function print out Dataframe that every path length of the data mapped.
            It's meaningful to digitize relationship.
         
         Parameters
@@ -916,8 +942,14 @@ class KeplerMapper(object):
 
         Returns
         -------
-        lens : Dataframe
+        result : Dataframe
             every path length of the data mapped.       
+        """
+        
+        
+        
+        result.index, result.columns = 
+        return result
             
             
             
