### 2019-07-15 b79f2cd
full values; just 0, 1, -1
I can full values only 0, 1. I'll think it more how can I obtain more than 2. (-1 is easy)
