### 2019-07-16 b9accec
final code; digitize_relationship
