+++ kmapper/kmapper.py
@@ -7,6 +7,7 @@ import itertools
 import os
 import sys
 import warnings
+import networkx as nx
 
 from jinja2 import Environment, FileSystemLoader, Template
 import numpy as np
@@ -901,61 +902,66 @@ class KeplerMapper(object):
     
     
     def digitize_relationship(
-            self,
-            graph,
-        ):
-        """Let members a, b each in nodes n, m. And define path length r of a and b; least number of edges that connect a, b. 
-           If a and b is connected by same node, then r is 0. And if a and b is not connected, then r is -1
-           This function print out Dataframe that every path length of the data mapped. It's meaningful to digitize relationship.
-
-        Parameters
-        ----------
-        graph : Dictionary
-            The data mapped
-
-        Returns
-        -------
-        result : Dataframe
-            every path length of the data mapped.       
-        """
-
-        # From graph, we can obtain nodes and links
-        nodes = graph.get('nodes')
-        nodes_k = list(nodes)
-        nodes_v = list(nodes.values())
-        links = graph.get('links')
-
-        # Made array for grouping linked node
-        relative = []
-        for node in nodes_k:
-            members = nodes.get(node)
-            linked = [members]
-            other_nodes = links.get(node)
-            if not other_nodes == None:
-                for other_node in other_nodes:
-                    linked.append(nodes.get(other_node))
-            relative.append(linked)
-
-        # Using a definition of path length, obtain the value between order pairs.
-        dict = {}
-        for linked in relative:
-            core_node = linked[0]
-            for node in linked:
-                for element in itertools.product(core_node, node):
-                    dict[element] = 1
-            for element in itertools.product(core_node, core_node):
-                    dict[element] = 0
+        self,
+        graph,
+        index,
+    ):
+    """Let members a, b each in nodes n, m. And define path length r of a and b; least number of edges that connect a, b. 
+       If a and b is connected by same node, then r is 0. And if a and b is not connected, then r is -1
+       This function print out Dataframe that every path length of the data mapped. It's meaningful to digitize relationship.
         
-        # It has to be solved how to obtaion path length more than 2...
-        
-        # Creating dataframe. Note we use the data index for index, columns of result.  
-        result = pd.Series(dict).unstack()
-        result = pd.DataFrame(result)
-        result.index, result.columns = Data.index, Data.index
+    Parameters
+    ----------
+    graph : Dictionary
+        The data mapped
+    index : list
+        Index of the data 
+    
+    Returns
+    -------
+    result : Dataframe
+        every path length of the data mapped.      
+     
+    """
 
-        # 'Nan' mean the order pairs be unrelated; path length is -1.
-        result.fillna(-1)
-        return result
-            
-            
-            
+    # From graph, we can obtain nodes and links
+    nodes = graph.get('nodes')
+    nodes_k = list(nodes)
+    nodes_v = list(nodes.values())
+    links = graph.get('links')
+    
+    # Made array for grouping linked node
+    relative = []
+    for node in nodes_k:
+        members = nodes.get(node)
+        linked = [members]
+        other_nodes = links.get(node)
+        if not other_nodes == None:
+            for other_node in other_nodes:
+                linked.append(nodes.get(other_node))
+        relative.append(linked)
+    
+    # Creating graph for using networkx module
+    G = nx.Graph()
+    for node in nodes_v:
+        nx.add_path(G, node)
+    
+    # Using a definition of path length, obtain the value between order pairs.
+    dict = {}
+    for member in range(len(index)):
+        for other_member in range(len(index)):
+            try:
+                dict[(member, other_member)] = nx.shortest_path_length(G, 
+                                                                       source=member, 
+                                                                       target=other_member, 
[…]
