### 2019-07-16 3adf400
Consider members in same node
