### 2019-07-16 5b0e17b
delete  core_node = linked[0]
