+++ kmapper/kmapper.py
@@ -962,7 +962,7 @@ class KeplerMapper(object):
         # Consider members in same node
         for linked in relative:
             for element in itertools.product(linked[0], linked[0]):
-                    dict[element] = 0
+                dict[element] = 0
 
         # Creating dataframe. Note we use the data index for index, columns of result.  
         result = pd.Series(dict).unstack()
