+++ kmapper/kmapper.py
@@ -902,11 +902,7 @@ class KeplerMapper(object):
 
         return X
     
-    def digitize_relationship(
-            self,
-            graph,
-            index
-        ):
+    def digitize_relationship(graph, index):
         """Let members a, b each in nodes n, m. And define path length r of a and b; least number of edges that connect a, b. 
            If a and b is connected by same node, then r is 0. And if a and b is not connected, then r is -1
            This function print out Dataframe that every path length of the data mapped. It's meaningful to digitize relationship.
@@ -916,7 +912,7 @@ class KeplerMapper(object):
         graph : Dictionary
             The data mapped
         index : list
-            Index of the data 
+            Index of original data 
 
         Returns
         -------
@@ -924,7 +920,6 @@ class KeplerMapper(object):
             every path length of the data mapped.      
 
         """
-
         # From graph, we can obtain nodes and links
         nodes = graph.get('nodes')
         nodes_k = list(nodes)
@@ -968,6 +963,5 @@ class KeplerMapper(object):
         result = pd.Series(dict).unstack()
         result = pd.DataFrame(result)
         result.index, result.columns = index, index
-
         return result
 
