### 2019-07-16 2bcb940
add 'self'
