### 2019-07-16 d1d7c2f
weight lightening
I deleted a few 'for phrases'.
