### 2019-07-16 bcc64ec
weight lightening
I deleted the unnecessary instructions.
