### 2019-07-21 0441857
remove default object constructors
