### 2019-07-21 229354c
fix tests to new api
