### 2019-07-21 #1315 @sauln
woah! sorry for stringing you along all this time, but it's done @leouieda! The mutable default arguments are now removed and we have a basic gallery set up to showcase examples.
