### 2019-07-22 0815059
Override Graphs to Apply Algorithms
