### 2019-08-31 #1315 @leouieda
Hi @labarba, I had a final look at the paper and everything seems good to go on my part. The author has addressed all of the major concerns of the reviewers.

@sauln I'm happy to move on with acceptance of your paper :tada: Here are the steps you need to take now:

- Make a release with the latest changes from the review and post the version number here.
- Archive the release on Zenodo/figshare and post the DOI here (make sure the title and author list matches the JOSS paper, if possible).
