### 2019-09-04 #1315 @sauln
I'll have some time finish off these last 3 things next week.

@deargle, @mlwave, and @smangham, could you please all confirm your affiliations and orcids?

- [ ] Make a release with the latest changes from the review and post the version number here.
- [ ] Archive the release on Zenodo/figshare and post the DOI here (make sure the title and author list matches the JOSS paper, if possible).
