### 2019-09-06 #1315 @deargle
I confirm my orcid and affiliation. Thanks for heading this, @sauln.

On Tue, Sep 3, 2019, 11:36 PM Nathaniel Saul <notifications@github.com>
wrote:

> I'll have some time finish off these last 3 things next week.
>
> @deargle <https://github.com/deargle>, @MLWave <https://github.com/MLWave>,
> and @smangham <https://github.com/smangham>, could you please all confirm
> your affiliations and orcids?
>
>    - Double check authors and affiliations (including ORCIDs)
>    - Make a release with the latest changes from the review and post the
>    version number here.
>    - Archive the release on Zenodo/figshare and post the DOI here (make
>    sure the title and author list matches the JOSS paper, if possible).
>
