### 2019-09-17 #1315 @smangham
I can confirm my name and affiliation, my ORCID is 0000-0001-7511-5652. Apologies for the delay and thanks for handling this @sauln!
