### 2019-09-25 4c51957
Update DOI in readme to project-wide one
The zenodo record service provides a DOI that will always resolve to the most recent version of kepler-mapper released there. It's a handy replacement so that the "cite" info in the readme does not need to constantly be updated along with new releases.
