### 2019-10-14 #1315 @deargle
Hi @leouieda, we published [an updated JOSS release to zenodo](https://zenodo.org/record/3485851), doi 10.5281/zenodo.3485851. Version number is 1.3.3 (took a few iterations to get the .zenodo.json formatting right)

Anything else you need from us?
