### 2019-10-14 #1315 @leouieda
@deargle looks good to me! I'm happy to pass this on for final acceptance into JOSS :tada:
