### 2019-10-14 #1315 @leouieda
@openjournals/joss-eics this submission is ready for acceptance and publication :confetti_ball:
