### 2019-10-15 #1315 @Kevin-Mattheus-Moerman
@sauln here are some minor points in relation to your paper.
- [ ] In "...data topoloigcal structures." fix the typo and consider rephrasing to "In "... topological data structures.""
- [ ] In "We leverage Scikit-Learn API-compatible cluster and scaling algorithms to construct network graphs in a flexible and user-friendly way.". Should you use "Scikit-Learn's" here?
- [ ] If works like matplotlib (which currently have hyperlinks) have a reference for a citation, please use/add them.
- [ ] I suggest the removal of the "Source code" section this the paper will contain a link to the repository already and the repository/readme should in term link to the documentation.

- [ ] The references `[@edelsbrunner2010computational], [@carlsson2009topology].` are comma separated, can you amend the paper.md file so this is a grouped citation? I.e. like this: [@edelsbrunner2010computational; @carlsson2009topology]
