### 2019-10-15 #1315 @Kevin-Mattheus-Moerman
@labarba can you help check the citation syntax for this paper? Are we happy about the use of `&` and `et al.` here?
