### 2019-10-15 #1315 @labarba
In-text citations are not properly used.

When the citation is part of the sentence, it renders: Author (year). If the citation is not part of the sentence, then it should appear in parenthesis. E.g. “We take inspiration from Smith et al. (2006) to blah “ and “The code implements Theory 1 (Jones, 2012) and Theory 2 (Roberts, 2014) to obtain blah “  For the syntax to obtain brackets in the right places, see https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax
