### 2019-10-17 1a94242
bibliography and citation fixes
