+++ paper.md
@@ -31,10 +31,10 @@ bibliography: paper.bib
 
 # Summary
 
-Topological data analysis (TDA) is a field analysis focused on understanding the shape and structure of complex data. By computing topological descriptors of data such as connected components, loops, and voids, we are better able to find hidden relationships among noisy and high-dimensionality data [@edelsbrunner2010computational], [@carlsson2009topology]. However, raw TDA metrics cannot be readily visualized. To address this gap, [@Singh2007] developed the Mapper algorithm in order to facilitate graphical exploration of data topoloigcal structures. The work of [@Lumetal2013] sparked widespread interest in the Mapper technique by demonstrating its use in multiple domains, including political science, biology, and sports analytics.
+Topological data analysis (TDA) is a field analysis focused on understanding the shape and structure of complex data. By computing topological descriptors of data such as connected components, loops, and voids, we are better able to find hidden relationships among noisy and high-dimensionality data [@edelsbrunner2010computational; @carlsson2009topology]. However, raw TDA metrics cannot be readily visualized. To address this gap, Singh et al. [-@Singh2007] developed the Mapper algorithm in order to facilitate graphical exploration of topological data structures. The work of Lum et al. [-@Lumetal2013] sparked widespread interest in the Mapper technique by demonstrating its use in multiple domains, including political science, biology, and sports analytics.
 
-This library, Kepler Mapper, is a Python implementation of the Mapper algorithm as first described in the paper “Topological Methods for the Analysis of High Dimensional Data Sets and 3D Object Recognition” [@Singh2007]. Kepler Mapper presents an intuitive interface for the Mapper algorithm along with multiple comprehension methods for visualizing the network graph that Mapper produces.
-We leverage Scikit-Learn API-compatible cluster and scaling algorithms to construct network graphs in a flexible and user-friendly way.
+This library, Kepler Mapper, is a Python implementation of the Mapper algorithm as first described in the paper "Topological Methods for the Analysis of High Dimensional Data Sets and 3D Object Recognition" [@Singh2007]. Kepler Mapper presents an intuitive interface for the Mapper algorithm along with multiple comprehension methods for visualizing the network graph that Mapper produces.
+We leverage Scikit-Learn-API-compatible clustering and scaling algorithms to construct network graphs in a flexible and user-friendly way.
 We also provide an extensive suite of tutorials detailing the use of Kepler Mapper for simple and complex use cases.
 
 ![Example Kepler Mapper graph visualization using the Wisconsin Breast Cancer Dataset [@Dua:2019]](http://i.imgur.com/ewjRodK.png)
@@ -44,12 +44,8 @@ We also provide an extensive suite of tutorials detailing the use of Kepler Mapp
 
 Kepler Mapper provides an object-oriented API for constructing lenses and building Mapper. The module employs the [strategy pattern](https://en.wikipedia.org/wiki/Strategy_pattern), giving users control over clustering algorithm, covering scheme, and nerve scheme. This allows the module to be flexible for many use cases. Clustering strategies follow the Scikit-Learn clusterer interface [@scikit-learn]. We provide similar interfaces for `Cover` classes and `Nerve` classes as well as default implementations of those classes that are most commonly found in the literature.
 
[…]
