### 2019-10-17 #1315 @deargle
@Kevin-Mattheus-Moerman I think I got everything from https://github.com/openjournals/joss-reviews/issues/1315#issuecomment-542142232. For point two, I changed the hyphenation to "Scikit-Learn-API-compatible ..."

@labarba I fixed the citation formatting.

Thinks look good to y'all?
