### 2019-10-17 #1315 @deargle
Would you want me to make a new release with those bibliography fixes?
