### 2019-10-17 #1315 @Kevin-Mattheus-Moerman
No that is okay. The main purpose is to archive the software. The paper can have minor changes.
