### 2019-10-17 #1315 @deargle
Thanks all!

Not a huge deal, but I noted that the suggested citation on JOSS has an extra period in it after the title. Fixable?

![image](https://user-images.githubusercontent.com/1174653/67045968-8ffa3b80-f0ec-11e9-8c5e-e7c98f8dbcfe.png)
