### 2019-10-17 #1315 @Kevin-Mattheus-Moerman
@openjournals/dev :point_up: can this be fixed?
